Flytt funksjoner ut i utilsfil
